Adding the ConnectivityStatusObserver, General House Keeping, Changing the Toast Charity dialog to a seperate screen

+++ app/src/main/java/com/ap/watchtogive/presentation/home/componants/DisplayCurrentCount.kt
@@ -1,18 +1,21 @@
 package com.ap.watchtogive.presentation.home.componants
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import java.text.NumberFormat
 import androidx.compose.ui.unit.sp
 import com.ap.watchtogive.presentation.theme.DarkTextColour
 import com.ap.watchtogive.presentation.theme.TextColour
+import java.text.NumberFormat
 
 @Composable
 fun DisplayCurrentCount(
-    currentTotal : Int
+    currentTotal : Long
 ) {
 
     val displayedTotal = NumberFormat.getInstance().format(currentTotal)
